A fix for problems with Men/Women in main navigation.
This fix adds classes to the main navigation and gender menu. With
thoses classes it is possible to hook into the main navigation click
event in browse.js. From there we can make it look like a click on Men
or Women in the main navigation is the same thing as a click in the
gender menu on the browse page.

This might not be the best solution, maybe a click in the main
navigation should reset the browse page.