Popular products in network now uses solr and display on browse page.
It is also possible to view popular products in network on the main
browse page through the 'f=1' querystring. A link is added to this
browse page from the popular products in your network sidebar.

This commit also fixes a problem with the brands list on the browse
page.

IMPORTANT: This update requires the search index to be rebuilt because
of changes made to schema.xml.