Updated how to handle sold out products during import.
When we run a feed, all products related to that feed is selected and
saved in a set. Every product found in the feed is then removed from
this set, and those that are left is counted as sold out products. For
every sold out product update the vendorproduct with the status sold out
(availability=0).

NOTE: This update changes both the db layout and the search index
layout. To apply restart solr and rebuild index.