Removed cache key product_detail2.
I removed this because the invalidate process was not working, that
cache region held looks, and if you edit a look you want it to
invalidate but we only have the look id not the product id which was
used for this cache region.

A possible solution is to check for this region in the invalidate map
and then go through every product and invalidate for them, if this
change seems to slow down the page we might need to investigate this
further.