Created my own auto_query based on the haystack version.
Searches in name, description (with 0.7 boost), manufacturer_name,
color_names and category_names. Supports keyword search with OR between
every keyword, sentence search with "" and negations with - infront of a
keyword.

This auto_query is only used in product searches, the default haystack
one is used for look and manufacturer searches.