Added override option to vendor categories in admin.
If override gender is set, all products for that vendor category will
have the new gender, regardless if it already had a gender set from the
feed. Default gender will continue to work as it should.

If override gender was set when it shouldn't you can use the reset
action for that vendor category to reset all products gender back to the
feed value.