Manufacturer search once again only searches for string matches.
Before this patched it show manufacturers that existed in products that
matched the search query, this is not a consistent behaviour with the
rest of the site.

This patch also removes the need to query the database when we list
manufacturers in browse / search. This patch changes the solr schema and
therefore needs to rebuild the index.