Update beta unlock page, added new landing page for selecting gender.
A new cookie is introduced called 'gender' it contains either 'M' or 'W'
and it set on the new landing page.

Beta middleware is moved to settings_common.py and must be called before
the select gender middleware.

An admin can add new background images for both the beta unlock page and
the new landing page for selecting gender, this is done through a new
model with only an image field.