Added ability to manually order categories.
This is solved by adding a new field to the category model called
name_order. The order is also dependent on the current language, this
was solved by translating the field with the modeltranslation module.

We continue to sort the categories in the client, but using a custum
data attribute on the a-element instead of the text.

Fixes #20774323