Introducing gender based urls that works without cookies.
This commit introduces a new templatetag called gender_url, it have the
same purpose as the built in url tag, except that this takes an
additional argument with the current gender and creates url based on
that.

This commit contains many search and replace changes to use this new
templatetag.

A new variable is always available in the request context, called
APPAREL_GENDER, it is set from the cookie, and gender specific views
override it in their own context.