Move most of the logic from django_facebook to profile.
We have no real usage for the middleware that comes with the
django_facebook module. We can ask for a facebook graph object if we
need it. We will also cache that facebook graph in our user session
until the access_token expires.

Login is handled by passing the access_token and uid from the facebook
javascript sdk to our own login view and use regular django login
methods.