Added ability to index and show discount prices.
Updated all importers to correctly import discount prices. If it is not
possible to find it in the feed, we instead match the imported price
with what is already stored for that product. If the imported price is
lower than the stored price we set the imported price as a discount
price. A negative aspect with this solution is the first price that is
imported will act as the base price and all other prices will be seen as
a discount. This only relates to feeds that does not have a discount
price field.

This commit requires a restart of solr for schema update, but should not
require a total rebuild of the index because it is only a new field
added and can safely be populated during feed imports.