Merged wardrobe with likes.
This commit contains a datamigration which will add all products in the
current wardrobe as likes and update them accordingly.

This is a hugh commit because it contains lots of search and replaces
for wardrobe to likes. A few files is still called wardrobe, this will
be fixed at a later time. As always watch out for errors caused by this
commit because of its size :)