Removed haystack.
We no longer use haystack. In a previously commit we removed haystack in
the searching proccess. This commit removes haystack from the data
mapping process also and haystack is therefor no longer in use.

By removing haystack from the mapping we can now add documents to solr
without commiting, which means that solr will auto commit in intervals
and in future solr commits this will result in near real-time search.
This also means that we no longer need to use celery, solr is fast
enough if no commit needs to be done.

This commit should increase the performance of the importer.