Updated importer api for images.
Instead of a single image-url the api now dictates that the field
image-url is a list of images with a specified minimum content length.
The images is then tried in order, if the first image is successfully
downloaded it will be used, if not the function continues with the next
image in the list. If no images could be downloaded the function will
raise SkipProduct exception.

All main providers are updated to support this api change.