Changed manufacturer to brand and made it possible to do mappings.
A brand mapping works just like a category mapping. Now you need to map both
a category and a brand to make it show up on browse.

The database migration might take awhile to run.

A new field was added to the product called static_brand, it is used to
together with sku to make sure it is possible to find a unique product.
Before the manufacturer field was used.