Added brand profiles.
This is a large commit. A foreign key for brand is created on the
apparel profile model. It is used to differentiates between a regular
profile and a brand profile.

A new task is created, the purpose of the task is to check every hour if
new products are added to a brand, if they are an actstream action is
created.

All brand links should now link to the brand profile instead of shop.