Updated django-activity-stream to version 0.4.3
To be able to upgrade django-activity-stream a large portion of the code
had to be changed. The main reason for this was that circular imports
happened due to new loading of action models in actstream (get_model on
models defined in settings).

A migration to update the actstream model is also included under the
profile app.

Moved translation to apparel apps and moved apparelrow.tasks to
apparel.task.