added published date to products
A published date is only set when a product is published for the first
time, it will never update. This will allow us to query for products
with published date null and published false. Those products has never
been shown on the site and is therefor safe to touch in the database.